+++ CHANGELOG.md
@@ -6,4 +6,5 @@
 - Remove main and dev branches, convert feature to main branch.
 - Rename previous feature branch to become main.
 - Remove html-validation checks due to template in-line styles being needed for optimal UI. 
-- Change workflows to update following major UI update.
+- Change workflows to update following major UI update. 
+- Update copy on resume to mirror resume